+++ src/types_impl/longdate.rs
@@ -1,5 +1,4 @@
 use crate::protocol::parts::hdb_value::HdbValue;
-use crate::protocol::parts::type_id::TypeId;
 use crate::{HdbError, HdbResult};
 use byteorder::{LittleEndian, ReadBytesExt};
 use chrono::{DateTime, Datelike, NaiveDate, NaiveDateTime, Timelike};
@@ -259,7 +258,7 @@ pub fn parse_longdate(nullable: bool, rdr: &mut io::BufRead) -> HdbResult<HdbVal
     let i = rdr.read_i64::<LittleEndian>()?;
     if i == NULL_REPRESENTATION {
         if nullable {
-            Ok(HdbValue::NULL(TypeId::LONGDATE))
+            Ok(HdbValue::NULL)
         } else {
             Err(HdbError::Impl(
                 "found NULL value for NOT NULL longdate column".to_owned(),